Fixed bug related to data type for storing memory usage.

+++ src/chart.hh
@@ -7,7 +7,8 @@
 #include <sstream>
 #include <algorithm>
 #include <vector>
-#include <map>
+#include <iomanip>
+#include <cmath>
 #include "util.hh"
 #include "../inc/NFont.h"
 
@@ -18,10 +19,10 @@ class Chart {
 public:
   Chart();
   void draw(NFont &font, NFont &fontSmall, SDL_Renderer *render, unsigned h);
-  void setPhysicalMemory(unsigned long long totalMemory, unsigned long long availMemory);
-  void addProcess(std::string name, unsigned long amount);
-  void setProcess(signed ix, unsigned long amount, bool add);
-  unsigned long getTotalUsage();
+  void setPhysicalMemory(size_t totalMemory, size_t availMemory);
+  void addProcess(std::string name, size_t amount);
+  void setProcess(signed ix, size_t amount, bool add);
+  size_t getTotalUsage();
   signed find(const std::string& processName);
   bool clear();
   unsigned getSize();
@@ -29,7 +30,7 @@ public:
 private:
   std::vector<Process> processList;
   std::map<std::string, SDL_Color> colorList;
-  unsigned long long totalMemory, usedMemory;
+  size_t totalMemory, usedMemory;
 };
 
 #endif
